ヘッダーの処理切り離し
