+++ app/(page)/(auth)/login/page.tsx
@@ -4,11 +4,16 @@ import Form from "@/app/_components/form";
 import { useLogin } from "@/app/_hocks/useLogin";
 
 import styles from "./LoginPage.module.css";
+import { useBase } from "@/app/_context/baseContext";
 
 // ログインページ
 const LoginPage = () => {
+  // 拠点取得
+  const {base} = useBase();
+  // 万が一拠点がnullだった場合
+  if(base === null) return;
   // ログインフック
-  const {register, handleSubmit, handleLogin, serverErrorMessages, errors, inputs} = useLogin();
+  const {register, handleSubmit, handleLogin, serverErrorMessages, errors, inputs} = useLogin({base: base});
 
   return(
     <div className={styles.login}>
